move package to presentation.common.notification
