+++ app/src/main/java/io/github/droidkaigi/confsched2018/presentation/common/notification/LocaleChangedBroadcastReceiver.kt
@@ -1,9 +1,8 @@
-package io.github.droidkaigi.confsched2018.presentation.common.broadcastreceiver
+package io.github.droidkaigi.confsched2018.presentation.common.notification
 
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import io.github.droidkaigi.confsched2018.util.NotificationHelper
 import timber.log.Timber
 
 class LocaleChangedBroadcastReceiver : BroadcastReceiver() {
