+++ app/src/main/java/io/github/droidkaigi/confsched2018/util/SessionAlarm.kt
@@ -6,7 +6,8 @@ import android.content.Context
 import android.os.Build
 import io.github.droidkaigi.confsched2018.R
 import io.github.droidkaigi.confsched2018.model.Session
-import io.github.droidkaigi.confsched2018.presentation.common.broadcastreceiver.NotificationBroadcastReceiver
+import io.github.droidkaigi.confsched2018.presentation.common.notification.NotificationBroadcastReceiver
+import io.github.droidkaigi.confsched2018.presentation.common.notification.NotificationContent
 import io.github.droidkaigi.confsched2018.util.ext.toReadableDateTimeString
 import java.util.concurrent.TimeUnit
 import javax.inject.Inject
